IGNITE-24770 Tests fix
