+++ modules/table/src/test/java/org/apache/ignite/internal/table/distributed/replication/PartitionReplicaListenerTest.java
@@ -306,8 +306,6 @@ public class PartitionReplicaListenerTest extends IgniteAbstractTest {
 
     private static final TablePartitionId commitPartitionId = new TablePartitionId(TABLE_ID, PART_ID);
 
-    private static final int ANOTHER_TABLE_ID = 2;
-
     private static final long ANY_ENLISTMENT_CONSISTENCY_TOKEN = 1L;
     private static final String TABLE_NAME = "test";
     private static final String TABLE_NAME_2 = "second_test";
@@ -744,62 +742,6 @@ public class PartitionReplicaListenerTest extends IgniteAbstractTest {
         completeBuiltIndexes(hashIndexStorage.storage(), sortedIndexStorage.storage());
     }
 
-    @Test
-    @WithSystemProperty(key = COLOCATION_FEATURE_FLAG, value = "false")
-    // TODO: IGNITE-24770 - remove this test after porting it to ZonePartitionReplicaListenerTest.
-    public void testTxStateReplicaRequestEmptyState() throws Exception {
-        doAnswer(invocation -> {
-            UUID txId = invocation.getArgument(5);
-
-            txManager.updateTxMeta(txId, old -> new TxStateMeta(
-                    ABORTED,
-                    localNode.id(),
-                    commitPartitionId,
-                    null,
-                    null,
-                    null
-            ));
-
-            return nullCompletedFuture();
-        }).when(txManager).finish(any(), any(), anyBoolean(), anyBoolean(), any(), any());
-
-        CompletableFuture<ReplicaResult> fut = partitionReplicaListener.invoke(TX_MESSAGES_FACTORY.txStateCommitPartitionRequest()
-                .groupId(tablePartitionIdMessage(grpId))
-                .txId(newTxId())
-                .enlistmentConsistencyToken(ANY_ENLISTMENT_CONSISTENCY_TOKEN)
-                .build(), UUID.randomUUID());
-
-        TransactionMeta txMeta = (TransactionMeta) fut.get(1, TimeUnit.SECONDS).result();
-
-        assertNotNull(txMeta);
-
-        assertEquals(ABORTED, txMeta.txState());
-    }
-
-    @Test
-    @WithSystemProperty(key = COLOCATION_FEATURE_FLAG, value = "false")
-    // TODO: IGNITE-24770 - remove this test after porting it to ZonePartitionReplicaListenerTest.
-    public void testTxStateReplicaRequestCommitState() throws Exception {
-        UUID txId = newTxId();
-
-        txStateStorage.putForRebalance(txId, new TxMeta(COMMITTED, singletonList(new EnlistedPartitionGroup(grpId)), clock.now()));
-
-        HybridTimestamp readTimestamp = clock.now();
-
-        CompletableFuture<ReplicaResult> fut = partitionReplicaListener.invoke(TX_MESSAGES_FACTORY.txStateCommitPartitionRequest()
-                .groupId(tablePartitionIdMessage(grpId))
-                .txId(txId)
-                .enlistmentConsistencyToken(ANY_ENLISTMENT_CONSISTENCY_TOKEN)
-                .build(), localNode.id());
-
-        TransactionMeta txMeta = (TransactionMeta) fut.get(1, TimeUnit.SECONDS).result();
-
-        assertNotNull(txMeta);
-        assertEquals(COMMITTED, txMeta.txState());
-        assertNotNull(txMeta.commitTimestamp());
-        assertTrue(readTimestamp.compareTo(txMeta.commitTimestamp()) > 0);
-    }
-
     @CartesianTest
     @CartesianTest.MethodFactory("finishedTxTypesFactory")
     void testExecuteRequestOnFinishedTx(TxState txState, RequestType requestType) {
@@ -814,38 +756,6 @@ public class PartitionReplicaListenerTest extends IgniteAbstractTest {
                 doSingleRowRequest(txId, testRow, requestType),
                 willThrowFast(TransactionException.class, "Transaction is already finished")
         );
-    }
-
-    @Test
-    public void testEnsureReplicaIsPrimaryThrowsPrimaryReplicaMissIfEnlistmentConsistencyTokenDoesNotMatchTheOneInLease() {
-        localLeader = false;
-
-        CompletableFuture<ReplicaResult> fut = partitionReplicaListener.invoke(TX_MESSAGES_FACTORY.txStateCommitPartitionRequest()
-                .groupId(tablePartitionIdMessage(grpId))
-                .txId(newTxId())
-                .enlistmentConsistencyToken(10L)
-                .build(), localNode.id());
-
-        assertThrowsWithCause(
-                () -> fut.get(1, TimeUnit.SECONDS).result(),
-                PrimaryReplicaMissException.class);
-    }
-
-    @Test
-    public void testEnsureReplicaIsPrimaryThrowsPrimaryReplicaMissIfNodeIdDoesNotMatchTheLeaseholder() {
-        localLeader = false;
-
-        placementDriver.setPrimaryReplicaSupplier(() -> new TestReplicaMetaImpl("node3", nodeId(3)));
-
-        CompletableFuture<ReplicaResult> fut = partitionReplicaListener.invoke(TX_MESSAGES_FACTORY.txStateCommitPartitionRequest()
-                .groupId(tablePartitionIdMessage(grpId))
-                .txId(newTxId())
-                .enlistmentConsistencyToken(ANY_ENLISTMENT_CONSISTENCY_TOKEN)
-                .build(), localNode.id());
-
-        assertThrowsWithCause(
-                () -> fut.get(1, TimeUnit.SECONDS).result(),
-                PrimaryReplicaMissException.class);
     }
 
     @Test
@@ -1553,70 +1463,6 @@ public class PartitionReplicaListenerTest extends IgniteAbstractTest {
         );
     }
 
-    @Test
-    @WithSystemProperty(key = COLOCATION_FEATURE_FLAG, value = "false")
-    // TODO: IGNITE-24770 - remove this test after porting it to ZonePartitionReplicaListenerTest.
-    public void testWriteIntentOnPrimaryReplicaSingleUpdate() {
-        UUID txId = newTxId();
-        AtomicInteger counter = new AtomicInteger();
-
-        testWriteIntentOnPrimaryReplica(
-                txId,
-                () -> {
-                    BinaryRow binaryRow = binaryRow(counter.getAndIncrement());
-
-                    return TABLE_MESSAGES_FACTORY.readWriteSingleRowReplicaRequest()
-                            .groupId(tablePartitionIdMessage(grpId))
-                            .tableId(TABLE_ID)
-                            .transactionId(txId)
-                            .requestType(RW_INSERT)
-                            .schemaVersion(binaryRow.schemaVersion())
-                            .binaryTuple(binaryRow.tupleSlice())
-                            .enlistmentConsistencyToken(ANY_ENLISTMENT_CONSISTENCY_TOKEN)
-                            .commitPartitionId(commitPartitionId())
-                            .coordinatorId(localNode.id())
-                            .timestamp(clock.now())
-                            .build();
-                },
-                () -> checkRowInMvStorage(binaryRow(0), true)
-        );
-
-        cleanup(txId);
-    }
-
-    @Test
-    @WithSystemProperty(key = COLOCATION_FEATURE_FLAG, value = "false")
-    // TODO: IGNITE-24770 - remove this test after porting it to ZonePartitionReplicaListenerTest.
-    public void testWriteIntentOnPrimaryReplicaUpdateAll() {
-        UUID txId = newTxId();
-        AtomicInteger counter = new AtomicInteger();
-
-        testWriteIntentOnPrimaryReplica(
-                txId,
-                () -> {
-                    int cntr = counter.getAndIncrement();
-                    BinaryRow binaryRow0 = binaryRow(cntr * 2);
-                    BinaryRow binaryRow1 = binaryRow(cntr * 2 + 1);
-
-                    return TABLE_MESSAGES_FACTORY.readWriteMultiRowReplicaRequest()
-                            .groupId(tablePartitionIdMessage(grpId))
-                            .tableId(TABLE_ID)
-                            .transactionId(txId)
-                            .requestType(RW_UPSERT_ALL)
-                            .schemaVersion(binaryRow0.schemaVersion())
-                            .binaryTuples(asList(binaryRow0.tupleSlice(), binaryRow1.tupleSlice()))
-                            .enlistmentConsistencyToken(ANY_ENLISTMENT_CONSISTENCY_TOKEN)
-                            .commitPartitionId(commitPartitionId())
-                            .coordinatorId(localNode.id())
-                            .timestamp(clock.now())
-                            .build();
-                },
-                () -> checkRowInMvStorage(binaryRow(0), true)
-        );
-
-        cleanup(txId);
-    }
-
     private void checkRowInMvStorage(BinaryRow binaryRow, boolean shouldBePresent) {
         Cursor<RowId> cursor = pkStorage().get(binaryRow);
 
@@ -1650,60 +1496,6 @@ public class PartitionReplicaListenerTest extends IgniteAbstractTest {
         }
     }
 
-    private void testWriteIntentOnPrimaryReplica(
-            UUID txId,
-            Supplier<ReadWriteReplicaRequest> updatingRequestSupplier,
-            Runnable checkAfterFirstOperation
-    ) {
-        partitionReplicaListener.invoke(updatingRequestSupplier.get(), localNode.id());
-        checkAfterFirstOperation.run();
-
-        // Check that cleanup request processing awaits all write requests.
-        CompletableFuture<UpdateCommandResult> writeFut = new CompletableFuture<>();
-
-        raftClientFutureClosure = cmd -> writeFut;
-
-        try {
-            CompletableFuture<ReplicaResult> replicaWriteFut = partitionReplicaListener.invoke(updatingRequestSupplier.get(),
-                    localNode.id());
-
-            assertTrue(replicaWriteFut.isDone());
-
-            raftClientFutureClosure = defaultMockRaftFutureClosure;
-
-            HybridTimestamp now = clock.now();
-
-            // Imitation of tx commit.
-            txStateStorage.putForRebalance(txId, new TxMeta(COMMITTED, new ArrayList<>(), now));
-            txManager.updateTxMeta(txId, old -> new TxStateMeta(
-                    COMMITTED, UUID.randomUUID(), commitPartitionId, now, null, null)
-            );
-
-            CompletableFuture<?> replicaCleanupFut = partitionReplicaListener.invoke(
-                    TX_MESSAGES_FACTORY.writeIntentSwitchReplicaRequest()
-                            .groupId(tablePartitionIdMessage(grpId))
-                            .tableIds(Set.of(grpId.tableId()))
-                            .txId(txId)
-                            .commit(true)
-                            .commitTimestamp(now)
-                            .build(),
-                    localNode.id()
-            );
-
-            assertFalse(replicaCleanupFut.isDone());
-
-            writeFut.complete(new UpdateCommandResult(true, true, 0));
-
-            assertThat(replicaCleanupFut, willSucceedFast());
-        } finally {
-            raftClientFutureClosure = defaultMockRaftFutureClosure;
-        }
-
-        // Check that one more write after cleanup is discarded.
-        CompletableFuture<?> writeAfterCleanupFuture = partitionReplicaListener.invoke(updatingRequestSupplier.get(), localNode.id());
-        assertThat(writeAfterCleanupFuture, willThrowFast(TransactionException.class));
-    }
-
     @Test
     void testWriteIntentBearsLastCommitTimestamp() {
         BinaryRow br1 = binaryRow(1);
@@ -1731,43 +1523,6 @@ public class PartitionReplicaListenerTest extends IgniteAbstractTest {
 
         UUID tx1 = newTxId();
         upsert(tx1, br1);
-    }
-
-    @ParameterizedTest
-    @ValueSource(booleans = {true, false})
-    @WithSystemProperty(key = COLOCATION_FEATURE_FLAG, value = "false")
-    // TODO: IGNITE-24770 - remove this test after porting it to ZonePartitionReplicaListenerTest.
-    void writeIntentSwitchForCompactedCatalogTimestampWorks(boolean commit) {
-        int earliestVersion = 999;
-        Catalog mockEarliestCatalog = mock(Catalog.class);
-        when(mockEarliestCatalog.version()).thenReturn(earliestVersion);
-
-        UUID txId = newTxId();
-        HybridTimestamp beginTs = beginTimestamp(txId);
-        HybridTimestamp commitTs = clock.now();
-
-        HybridTimestamp reliableCatalogVersionTs = commit ? commitTs : beginTs;
-        when(catalogService.activeCatalog(reliableCatalogVersionTs.longValue())).thenThrow(new CatalogNotFoundException("Oops"));
-        when(catalogService.earliestCatalog()).thenReturn(mockEarliestCatalog);
-
-        CompletableFuture<ReplicaResult> invokeFuture = partitionReplicaListener.invoke(
-                    TX_MESSAGES_FACTORY.writeIntentSwitchReplicaRequest()
-                            .groupId(tablePartitionIdMessage(grpId))
-                            .tableIds(Set.of(grpId.tableId()))
-                            .txId(txId)
-                            .commit(commit)
-                            .commitTimestamp(commit ? commitTs : null)
-                            .build(),
-                    localNode.id()
-            );
-
-        assertThat(invokeFuture, willCompleteSuccessfully());
-        assertThat(invokeFuture.join().applyResult().replicationFuture(), willCompleteSuccessfully());
-
-        verify(mockRaftClient).run(commandCaptor.capture());
-        WriteIntentSwitchCommand command = (WriteIntentSwitchCommand) commandCaptor.getValue();
-
-        assertThat(command.requiredCatalogVersion(), is(earliestVersion));
     }
 
     /**
@@ -1868,68 +1623,8 @@ public class PartitionReplicaListenerTest extends IgniteAbstractTest {
         cleanup(tx1);
     }
 
-    @Test
-    @WithSystemProperty(key = COLOCATION_FEATURE_FLAG, value = "false")
-    // TODO: IGNITE-24770 - remove this test after porting it to ZonePartitionReplicaListenerTest.
-    public void abortsSuccessfully() {
-        AtomicReference<Boolean> committed = interceptFinishTxCommand();
-
-        CompletableFuture<?> future = beginAndAbortTx();
-
-        assertThat(future, willSucceedFast());
-
-        assertThat(committed.get(), is(false));
-    }
-
-    private CompletableFuture<?> beginAndAbortTx() {
-        when(txManager.cleanup(any(), any(Map.class), anyBoolean(), any(), any())).thenReturn(nullCompletedFuture());
-
-        HybridTimestamp beginTimestamp = clock.now();
-        UUID txId = transactionIdFor(beginTimestamp);
-
-        TxFinishReplicaRequest commitRequest = TX_MESSAGES_FACTORY.txFinishReplicaRequest()
-                .groupId(tablePartitionIdMessage(grpId))
-                .commitPartitionId(tablePartitionIdMessage(grpId))
-                .txId(txId)
-                .groups(Map.of(tablePartitionIdMessage(grpId), partitionEnlistmentMessage(localNode.name(), Set.of(grpId.tableId()))))
-                .commit(false)
-                .enlistmentConsistencyToken(ANY_ENLISTMENT_CONSISTENCY_TOKEN)
-                .build();
-
-        return partitionReplicaListener.invoke(commitRequest, localNode.id());
-    }
-
-    private PartitionEnlistmentMessage partitionEnlistmentMessage(String primaryConsistentId, Set<Integer> tableIds) {
-        return TX_MESSAGES_FACTORY.partitionEnlistmentMessage()
-                .primaryConsistentId(primaryConsistentId)
-                .tableIds(tableIds)
-                .build();
-    }
-
     private static UUID transactionIdFor(HybridTimestamp beginTimestamp) {
         return TestTransactionIds.TRANSACTION_ID_GENERATOR.transactionIdFor(beginTimestamp);
-    }
-
-    @Test
-    @WithSystemProperty(key = COLOCATION_FEATURE_FLAG, value = "false")
-    // TODO: IGNITE-24770 - remove this test after porting it to ZonePartitionReplicaListenerTest.
-    public void commitsOnSameSchemaSuccessfully() {
-        when(validationSchemasSource.tableSchemaVersionsBetween(anyInt(), any(), any(HybridTimestamp.class)))
-                .thenReturn(List.of(
-                        tableSchema(CURRENT_SCHEMA_VERSION, List.of(nullableColumn("col")))
-                ));
-
-        AtomicReference<Boolean> committed = interceptFinishTxCommand();
-
-        CompletableFuture<?> future = beginAndCommitTx();
-
-        assertThat(future, willSucceedFast());
-
-        assertThat(committed.get(), is(true));
-    }
-
-    private static CatalogTableColumnDescriptor nullableColumn(String colName) {
-        return new CatalogTableColumnDescriptor(colName, ColumnType.INT32, true, 0, 0, 0, DefaultValue.constant(null));
     }
 
     private static CatalogTableColumnDescriptor defaultedColumn(String colName, int defaultValue) {
@@ -1951,70 +1646,6 @@ public class PartitionReplicaListenerTest extends IgniteAbstractTest {
         };
 
         return committed;
-    }
-
-    private CompletableFuture<?> beginAndCommitTx() {
-        when(txManager.cleanup(any(), any(Map.class), anyBoolean(), any(), any())).thenReturn(nullCompletedFuture());
-
-        HybridTimestamp beginTimestamp = clock.now();
-        UUID txId = transactionIdFor(beginTimestamp);
-
-        HybridTimestamp commitTimestamp = clock.now();
-
-        TxFinishReplicaRequest commitRequest = TX_MESSAGES_FACTORY.txFinishReplicaRequest()
-                .groupId(tablePartitionIdMessage(grpId))
-                .commitPartitionId(tablePartitionIdMessage(grpId))
-                .txId(txId)
-                .groups(Map.of(tablePartitionIdMessage(grpId), partitionEnlistmentMessage(localNode.name(), Set.of(grpId.tableId()))))
-                .commit(true)
-                .commitTimestamp(commitTimestamp)
-                .enlistmentConsistencyToken(ANY_ENLISTMENT_CONSISTENCY_TOKEN)
-                .build();
-
-        return partitionReplicaListener.invoke(commitRequest, localNode.id());
-    }
-
-    @Test
-    @WithSystemProperty(key = COLOCATION_FEATURE_FLAG, value = "false")
-    // TODO: IGNITE-24770 - remove this test after porting it to ZonePartitionReplicaListenerTest.
-    public void commitsOnCompatibleSchemaChangeSuccessfully() {
-        when(validationSchemasSource.tableSchemaVersionsBetween(anyInt(), any(), any(HybridTimestamp.class)))
-                .thenReturn(List.of(
-                        tableSchema(CURRENT_SCHEMA_VERSION, List.of(nullableColumn("col1"))),
-                        // Addition of a nullable column is forward-compatible.
-                        tableSchema(FUTURE_SCHEMA_VERSION, List.of(nullableColumn("col1"), nullableColumn("col2")))
-                ));
-
-        AtomicReference<Boolean> committed = interceptFinishTxCommand();
-
-        CompletableFuture<?> future = beginAndCommitTx();
-
-        assertThat(future, willSucceedFast());
-
-        assertThat(committed.get(), is(true));
-    }
-
-    @Test
-    @WithSystemProperty(key = COLOCATION_FEATURE_FLAG, value = "false")
-    // TODO: IGNITE-24770 - remove this test after porting it to ZonePartitionReplicaListenerTest.
-    public void abortsCommitOnIncompatibleSchema() {
-        simulateForwardIncompatibleSchemaChange(CURRENT_SCHEMA_VERSION, FUTURE_SCHEMA_VERSION);
-
-        AtomicReference<Boolean> committed = interceptFinishTxCommand();
-
-        CompletableFuture<?> future = beginAndCommitTx();
-
-        IncompatibleSchemaAbortException ex = assertWillThrowFast(future, IncompatibleSchemaAbortException.class);
-
-        assertThat(ex.getMessage(), containsString("Commit failed because schema is not forward-compatible [fromSchemaVersion=1, "
-                + "toSchemaVersion=2, table=test, details=Column default value changed]"));
-
-        assertThat(committed.get(), is(false));
-    }
-
-    private void simulateForwardIncompatibleSchemaChange(int fromSchemaVersion, int toSchemaVersion) {
-        when(validationSchemasSource.tableSchemaVersionsBetween(anyInt(), any(), any(HybridTimestamp.class)))
-                .thenReturn(incompatibleSchemaVersions(fromSchemaVersion, toSchemaVersion));
     }
 
     private void simulateBackwardIncompatibleSchemaChange(int fromSchemaVersion, int toSchemaVersion) {
@@ -2720,69 +2351,6 @@ public class PartitionReplicaListenerTest extends IgniteAbstractTest {
         testRoOperationFailsIfTableWasDropped(onExistingRow, (targetTxId, readTimestamp, key) -> {
             return doRoScanRetrieveBatchRequest(targetTxId, readTimestamp);
         });
-    }
-
-    @Test
-    @WithSystemProperty(key = COLOCATION_FEATURE_FLAG, value = "false")
-    // TODO: IGNITE-24770 - remove this test after porting it to ZonePartitionReplicaListenerTest.
-    void commitRequestFailsIfCommitPartitionTableWasDropped() {
-        testCommitRequestIfTableWasDropped(grpId, Map.of(grpId, localNode.name()), grpId.tableId());
-    }
-
-    @Test
-    @WithSystemProperty(key = COLOCATION_FEATURE_FLAG, value = "false")
-    // TODO: IGNITE-24770 - remove this test after porting it to ZonePartitionReplicaListenerTest.
-    void commitRequestFailsIfNonCommitPartitionTableWasDropped() {
-        TablePartitionId anotherPartitionId = new TablePartitionId(ANOTHER_TABLE_ID, 0);
-
-        testCommitRequestIfTableWasDropped(grpId, Map.of(grpId, localNode.name(), anotherPartitionId, localNode.name()),
-                anotherPartitionId.tableId());
-    }
-
-    private void testCommitRequestIfTableWasDropped(
-            TablePartitionId commitPartitionId,
-            Map<TablePartitionId, String> groups,
-            int tableToBeDroppedId
-    ) {
-        when(validationSchemasSource.tableSchemaVersionsBetween(anyInt(), any(), any(HybridTimestamp.class)))
-                .thenReturn(List.of(
-                        tableSchema(CURRENT_SCHEMA_VERSION, List.of(nullableColumn("col")))
-                ));
-        when(txManager.cleanup(any(), any(Map.class), anyBoolean(), any(), any())).thenReturn(nullCompletedFuture());
-
-        AtomicReference<Boolean> committed = interceptFinishTxCommand();
-
-        UUID txId = newTxId();
-        HybridTimestamp txBeginTs = beginTimestamp(txId);
-
-        String tableNameToBeDropped = catalogService.activeCatalog(txBeginTs.longValue()).table(tableToBeDroppedId).name();
-
-        makeTableBeDroppedAfter(txBeginTs, tableToBeDroppedId);
-
-        CompletableFuture<?> future = partitionReplicaListener.invoke(
-                TX_MESSAGES_FACTORY.txFinishReplicaRequest()
-                        .groupId(tablePartitionIdMessage(commitPartitionId))
-                        .commitPartitionId(tablePartitionIdMessage(commitPartitionId))
-                        .groups(
-                                groups.entrySet().stream()
-                                        .collect(toMap(
-                                                e -> tablePartitionIdMessage(e.getKey()),
-                                                entry -> partitionEnlistmentMessage(entry.getValue(), Set.of(entry.getKey().tableId()))
-                                        ))
-                        )
-                        .txId(txId)
-                        .enlistmentConsistencyToken(ANY_ENLISTMENT_CONSISTENCY_TOKEN)
-                        .commit(true)
-                        .commitTimestamp(clock.now())
-                        .build(),
-                localNode.id()
-        );
-
-        IncompatibleSchemaAbortException ex = assertWillThrowFast(future, IncompatibleSchemaAbortException.class);
-
-        assertThat(ex.getMessage(), is("Commit failed because a table was already dropped [table=" + tableNameToBeDropped + "]"));
-
-        assertThat("The transaction must have been aborted", committed.get(), is(false));
     }
 
     @CartesianTest
